refactor(reports): extract meeting detection to reports_meetings.go
